SX publish api action: Restore handling of unset newrevid
In some rare cases (e.g. when publishing already present sections
with the exact same contents as the current revision), the "newrevid"
index of the edit result can be unset. In these cases, the CX/SX tags
cannot be added. This was previously fixed but a regression issue led
this fix to be lost. This patch restores the fix and also updates the
translation inside "cx_translations" table only when "newrevid" is
set.

Change-Id: I384fb03662bb1a7c096e8812b90041061eaf7201