SX MT validation: fix calls of "calculateUnmodifiedContent" method
The mtHelper/calculateUnmodifiedContent method expects a language
parameter, so that it can use a slightly different approach for CJK
languages.

However, the "calculateScore" and "getMTScoreForPageSection" methods,
inside mtValidator module, that use the "calculateUnmodifiedContent"
method, do not pass any language parameter. That leads to issues
when calculating the unmodified content percentage for CJK languages.

This patch fixes this issue by adding a language parameter to the
"calculateScore" and "getMTScoreForPageSection" methods, which they
forward to the "calculateUnmodifiedContent" method. All calls of
these methods are also fixed to pass this parameter.

Change-Id: I0d2b2ad2b06b1bba1ef0332b7d9897fa3cc86457