SX: Remove "publishFeedbackMessages" from vuex state
Publish feedback messages are related only to the publishing
functionality and are not shared between different steps of
the application. Until now, they lived inside the application
state, which they clearly do not belong to.

This patch fixes this problematic situation by removing the
"publishFeedbackMessages" application state variable and adding
it as ref property inside "usePublishTranslation" composable,
to be used by the interested parts of the "Preview and publish"
step. All related mutations and getters are removed, while the
"validateMT" and "publishTranslation" translator actions are
modified to return a publish feedback message (or null) instead
of adding it to the application state variable.

Tests are fixed too.

Change-Id: I1a0155de6551323c5454710f292414cb1b1e5691