SX parallel corpora: Fix parallelCorporaMTContent calculation
Currently, the calculation of the content for the parallel corpora
MT unit, takes into account all sentences of a subsection (for
subsections that are not block templates), when checking if the
same MT provider is used for the whole subsection. However, only
translated sentences should be taken into account, since we do
not care for untranslated sentences within a subsection.

This patch fixes this issue by filtering out untranslated
sentences when checking if the same MT provider is used across
the whole subsection.

Bug: T270499
Change-Id: I1b10e616111e13bf5a5ad0ee5dbc234e7fb34230