SX: Add CAPTCHA confirmation dialog before publishing, when needed
This patch adds an initial version of the Captcha confirmation dialog
that should be displayed to the users for some target wikis, before
publishing. This is only a first implementation and some refinement
will be required for the dialog to be ready for production.

Tested with "SimpleCaptcha", "QuestyCaptcha" and "FancyCaptcha".

Bug: T272586
Change-Id: I06665027967c4865c084317e9b36f36f1c481823