Section translation Desktop editor: Load a section
If the URL contains sourcesection param, that particular section
alone is shown in editor.

Also contains modifications to data models to have source and target
section titles. And removal of some old code related to loading single
section that was not used or functional.

Category UI will be hidden in section translation mode.

Bug: T311614
Change-Id: Ie2b5fbfae75242ae6955ee9d1008e39b6e750a41