SX new by translation mobile entrypoint: Hide invite when user types
Until now the invite for the "new-by-translation" mobile entrypoint
is being hidden when user focuses on the editing surface. However,
in order to provide a better UX we want to hide this invite only
when user types inside the editing surface.

This patch fixes this issue by replacing the "focus" event listener
with an "input" event listener.

Bug: T298084
Change-Id: Ie2b72296482f86fadd64a2e196574b0e906c5ee4