SX in Desktop: fix error on editing enablement after publishing
"mw.cx.TranslationController.enableEditing" method is called when
publishing is completed (either successful or failed). This method
re-enables the category UI of the translation view. However, this
"categoryUI" variable is null for section translations, leading to
a JS error.

This patch fixes this issue, by checking if the "categoryUI" variable
is set before calling "disableCategoryUI" method on it.

Bug: T311997
Change-Id: I971c8ca33b1c070e24e7759d136140e250d17c13