Move section position calculation logic to server-side
Until now the calculation of the section position is handled by
Section Translation client side code. In order to have the same
functionality for desktop section translation, it is better to
move this functionality inside "ApiSectionTranslationPublish"
API action that is being called by both desktop and mobile section
translation.

To fulfil that need, this patch removes the `sectionnumber` parameter
from "cxpublishsection" API action and adds a new service named
"SectionPositionCalculator" that calculates the section position. This
service is used by the API action.

This patch also removes all the code related to the section calculation
logic from Section Translation codebase and fixes the save and publish
payloads for both desktop and mobile versions.

Bug: T311997
Change-Id: I29738929199f1a5964d132cca7a988d6de63252a