SX section positioning: Use "prependtext" param in "edit" API request
Until now, proper section positioning is implemented by replacing the
contents of the first appendix section with a concatenation of the
new section contents with the first appendix section contents. This
requires an API request to fetch the contents of the first appendix
section. Additionally, the first appendix section often contains
references which are difficult to be resolved properly, leading to
broken references after section publishing in a lot of cases.

To solve this issue, this patch proposes the usage of the
"prependtext" parameter inside the payload of the Edit API action,
that is being called by the "cxpublishsection" API. This parameter
tells the Edit API to add this text to the beginning of the section,
so there is no more the need to fetch and concatenate the appendix
section contents.

Since it's not anymore needed, this patch also removes the
"SectionContentFetcher" service.

Bug: T311997
Change-Id: I9b7967727f07987bc743f89338ec531084b1cd9e