SX: Move section title prepending to server-side
Because of Edit API limitations, when a section needs to be
positioned before the first appendix section of the target
article, we also need to prepend the target section title
to the HTML contents that are being published.

This functionality is currently, only implemented inside
Section Translation codebase. In order to be also used
by desktop section translations, this functionality is
better to be moved to the server-side code that is being
used by both the mobile and the desktop version.

Moreover, inside Section Translation application, the target
section title is being prepended in all cases, even in the
cases when it shouldn't - the section title should not be
prepended for lead sections, sandbox sections and when appendix
section titles are missing from the target article.

This patch moves the section title prepending functionality
to the server side, removes it from Section Translation codebase,
and only applies it for the required case (when appendix section
titles exist inside target article).

Bug: T311997
Bug: T316761
Bug: T316881
Change-Id: I75a3e92faa7500a3920a2e8b2cb2cad5275a5ee3