SX: Fix the message explaining the expected publishing result
The message explaining the expected publishing result, based on the
selected publish target, doesn't change when the publish target is
updated.

This patch fixes this issue and also enforces the message for
sandbox sections to be always shown when publish target is
"SANDBOX_SECTION", even when translating a lead section (which
was not the case before).

Bug: T316933
Change-Id: I5f2ad71abddbabe3499edded6ddac08783079922