SX Instrumentation: Fix event logging access_method
Until now the "access_method" parameter for the event logging payload
is populated based on the value of "skin" global configuration
parameter (mw.config.get("skin")). That is, when the "skin" parameter
is equal to "minerva", the access_method is set to "mobile web",
otherwise it's set to "desktop". However, the skin is always
"contenttranslation" for both desktop and mobile CX, and that causes
the "access_method" parameter to always be set to "desktop".

This patch fixes this issue inside SX codebase, by hardcoding the
value of "access_method" parameter to "mobile web", since this code
is only executed for mobile versions.

Bug: T295757
Change-Id: Ib77c1751d2ab3e101301015b4bf4e1d9a70ad73a