SX: Move target title making for sandbox sections to server-side
This patch moves the construction of the target page title for
sandbox target to the ApiSectionTranslationPublish action, so
that this logic can be used by both desktop and mobile section
translations. Additionally, it removes all the related code from
the client-side codebase.

Since the target title used for publishing is no longer available
in client-side, this patch also adds a "targettitle" field in the
result returned by the API, so that it can be used in client-side
for proper redirection to the target page, after successful
publishing.

Tests are fixed too.

Bug: T275363
Change-Id: I811e12d691966561fcaa8872ddc0f0de1b19862e