Desktop SX: Hide section title placeholder inside VE for target page
Currently, section title is hidden properly inside VE for source
page, however there is still a placeholder for translating the
section title inside target page VE. This is because we only hide
section nodes which are H2. Since, this is never the case for target
section nodes (which are always placeholder nodes), we should check
if the corresponding source section node is H2, to hide both source
and target nodes for section titles.

Bug: T316929
Change-Id: I12db19fde4ee7d6395e5bc989fb5508c3cc60ce6