CX: Use suggested title from cxserver as target title when no URL param
This patch adds the "getSuggestedTitle" method inside the machine
translation service. This method sends a request to cxserver, to
fetch the suggested target title and returns it.

The "getSuggestedTitle" method is used before translation initialization
to get a suggested target title, when no "targettitle" URL param exists.

Additionally, this patch removes "targettitle" URL param from the
constructed URL inside "startPageInCX" handler, when the URL param
is not explicitly provided in the given URL.

Bug: T225494
Change-Id: I0d4fbff970af42dbc169d8974256f5f90c08367b