TranslationCorporaStore: Avoid unnecessary lock when record does not exist
Doing 'FOR UPDATE' for non-existing records will lock the whole gap.
This is undesirable. We need to lock the row only when it exists in cx_corpora
table.

For fixing this, run the doFind method on replica. If the record exists,
run doFind again with primary db with 'FOR UPDATE'

The transactions under doAtomicSection were reduced by keeping '..for update'
and update and moving all other logic to outside.

Bug: T256229
Change-Id: I1ade0de6f02f958b31c74a31f6f804a46deef7ac