SX Block templates support: Add mapping indicator
This patch adds a mapping indicator as described in the related task.
This indicator is basically an appropriate icon nested inside a
circular progress bar, which indicates the percentage of source
template parameters that were successfully adapted inside the
target template. The indicator is implemented by the
"BlockTemplateStatusIndicator" SFC.

To support the implementation, some getters and methods are also
added inside subSection model, to provide the necessary data.

Bug: T294650
Change-Id: I4f3fdb5cb5128b4e74e5cc83f67a325c5a6e6e27