Fix SiteMapper::getCXServerURL to also handle protocol relative paths
Section positioning in SX doesn't work properly for languages that do
not have their appendix section titles hardcoded (i.e. all languages
except "en", "bn", "es", "fr" and "de"). That happens because the
request to fetch the appendix section titles for the target language
from the cxserver fails. The reason of this failure is that the
"getCXServerURL" method inside SiteMapper, expects the
"$wgContentTranslationSiteTemplates['cx']" configuration to be
a URL containing the "https" protocol, which is not the case for
production wikis ("$wgContentTranslationSiteTemplates['cx']" is set
to "//cxserver.wikimedia.org/v1" inside wmf-config/CommonSettings.php).

To fix this issue, this patch modifies the "getCXServerURL" method to
also handle protocol relative paths inside
"$wgContentTranslationSiteTemplates['cx']" configuration.

Bug: T321671
Change-Id: I8c2c791a4680cdaba7d38d796b69962873a664ce