SX: Fix redirection after successful publishing for non-ASCII titles
The "targetTitle" that is returned by the "cxpublishsection" endpoint,
is in a URL-encoded form. For this reason, this target title should be
decoded before passed as an argument to the "mw.util.getUrl" method,
otherwise the final URL is malformed and leads to a "Bad title"
exception, after redirection upon successful publishing.

Bug: T321665
Change-Id: Iabd9b063000f3271d71742a2f5092ec14e909b98