SX: Fix publishTranslation api method returned object on api error
The "publishTranslation" api method is expected to always return a
promise that resolves to an object containing a "targetTitle" and a
"publishFeedbackMessage" property, that is an object is the following
form: {publishFeedbackMessage, targetTitle }. However, this is not
happening when an API error (other than a CAPTCHA error) occurs. In
these cases, the resolved object contains a "publishMessage" property
leading to broken error handling.

This patch fixes this issue by returning the proper object inside
the "publishTranslation" API method.

Change-Id: I10c13899de2f8186d2f59223c7185186d77237c3