contributionsmenu entrypoint: Use unique ids for portlet links in sticky header
Change the id of links for sticky header so that they are unique.
Since the vector skin will use id attributes for constructing icons,
we need to supply icon for this new id. Since the icons are coming from
the RL module ext.cx.icons, add a sticky variant too.

That sticky variant is not really a variant but a copy of original icons.
So, it is hack we have to add since icons cannot be passed but constructed
programmatically by Vector

Bug: T320448
Change-Id: Ic51f56b1720da24ad0abc2276a726a6e9b2cff3f