SX MW UI lib: Fix MWThumbnail defaults and usage
Currently, in all cases where MWThumbnail UI component is used, it is
passed a nonexistent "width" property and then it is styled in CSS so
that it has the proper size (width, height).

Additionally, some inconsistencies between the current implementation
and the design specification has been observe. More specifically:
1. The placeholder icon used when no valid thumbnail is provided, is
the "imageLayoutFrameless" icon instead of the "article" icon.
2. The size of the placeholder icon is set to 80px. According to the
design, the size of the placeholder icon itself should be set to 50px
and the size of the placeholder including the background should be
equal to 84px, which is the thumbnail size.
3. The color of the placeholder icon is currently unset, which leads
base10 (#202122) to finally be used, while it should be the base20.
4. The background color of the placeholder icon is currently set to
the base90 color, while the base80 should be used instead.

This patch adds new props to the MWThumbnail and fixes the usages of
this component, to remove the nonexistent and not used "width" prop
and fix the above inconsistencies.

Bug: T253253
Change-Id: Id1e1c08f3d2966439a35cf8953df5c6860eed2be