SX: Fix block template edit to use translated content if exists
Currently, when user tries to edit an adapted block template,
the proposed translation is used, even if the template has
already been edited. That leads to lost edits, when user tries
to re-edit the template.

To fix this issue, this patch uses the proper payload for the
SX editor route, when user clicks on the "Review and edit the
translation" button.

Bug: T323818
Change-Id: I4aa71a711ba82e2121e1a825cc28731853431f89