SX: Supported language codes for language selector
This change allows the target language to be listed among the ones
available in the article search. As a result, this would enable the
target language to be used as the source language by swiching the
two (target language becomes source language) and not break the UX flow
but provide an alternative to translating between languages by changing
between source and target selections.

Bug: T293387
Change-Id: I72ca93720f39bc5cf4d1db10b7180dfa88c70736