SX: Add auto-save after each segment translation
This patch adds a "saveTranslation" vuex action that sends a request
to the "sxsave" with the proper payload to store the parallel corpora
and updates the corresponding rows in the "cx_translations" and
"cx_section_translations" db tables.

This action is dispatched after each segment translation, i.e. after
user clicks on the "Apply the translation" button, or after the edit
of the proposed translation of the segment. These calls are debounced
so that the action is executed at most once every 3 seconds.

Finally, this action is also dispatched before successful translation
publishing.

Bug: T322517
Change-Id: I775f46651746fc52d573d02b8f0132508df40f3c