SX: Add findSectionTranslationsByUser method and SectionTranslationDTO
This patch adds the "findSectionTranslationsByUser" method inside the
"SectionTranslationStore" service to support fetching (draft) section
translations.

Additionally, this patch introduces the SectionTranslationDTO class
that will be used to encapsulate the necessary payload for section
translation retrieval requests. Unlike the Entity\SectionTranslation
class that is mapped to the "cx_section_translations" database table,
this DTO class is tuned for exporting section translation data to
views.

Bug: T322517
Change-Id: I4178d125f1ac59bd513932efd1cdb77dc2410a2d