SX: Add "currentTranslation" application state variable
This patch introduces the "currentTranslation" variable inside
application state, to store the draft section translation that
is selected to be continued.

Until now, the only "translatable" entities were SectionSuggestion
models. For that reason the "currentSectionSuggestion" state variable
is used in several places across the application, as the main piece
of the translation process, and its properties are often used as
source of truth.

However, once the support for draft translations is completed, users
will be able to continue a previously saved draft translation, thus
there will be no suggestion selected for these cases, but rather an
"active" Translation model. For that reason, the usage of the
"currentSectionSuggestion" variable is replaced in all places where
it's not necessary - this variable is needed during the "Confirm a
translation", "Pick a section" and "Compare the contents" steps,
since their functionality revolves around the currently selected
suggestion. An example of such replacement is the replacement of the
usage of "currentSectionSuggestion.sourceTitle" with the title of
the current source page ("getters.getCurrentPage.title").

Tests have also been updated properly.

Change-Id: I3cd3c73a2766bf5f08e7064129bb34d14ea2e2eb