CX: Introduce TranslationUnitDTO class and use it where needed
The data stored inside "cx_corpora" have to be properly formatted
before returned to the "client" of these data. The form is which
the data should be transformed is different for each usecase.

In order to provide a unified way of handling these data, and better
encapsulate them inside a single class, this patch introduces the
"TranslationUnitDTO" class. Inside this class, there have been moved
all the related data-mangling methods that are applied on the corpora
data to transform them in the desired form.

Bug: T322521
Change-Id: I6029ce9580555ee4d30f0d000d31e194a1d1181e