SX: Basic support for leaving an auto-saving translation on mobile
Until now, once a translation has been started (at least one segment
has been translated), and the user is trying to leave the page, a
warning message is being displayed. To support this functionality,
a "translationInProgress" state variable holds a boolean value that
indicates whether the message should be displayed on page unload.

However, since the auto-saving functionality for translations has
been deployed, there is no need to display a warning about losing
the draft translation, to the users, on every occasion. The only case
to display such a warning is when an "auto-save" request is pending
(that is it's debounced and not yet executed) or an "auto-save" HTTP
request has been sent but not yet completed.

To support this functionality, two new state variables have been added
to the application state, so that we can easily check whether the
warning should be displayed to the users.

Bug: T322518
Change-Id: I107e1e20b64427f81317f327a987d5dd64d3e24a