Test ServiceWiring.php and other fixes to tests
Test ServiceWiring.php using tests copied from CentralAuth. Because
phpunit does not support marking a file as covered, the ServiceWiring
file is ignored for code coverage as the tests fully cover the file.

Add TestAutoloadNamespaces to extension.json which will autoload
phpunit test files for this extension. Fix namespace for the
AbuseFilterCheckerTest.php file to be consistent with the other
tests as well as adding the group for the extension to other test
files.

Change-Id: Ibe32f86ffe44d9491ee44d22e2db58142cd30472