Fix publishing target for sandbox section translations on desktop
"cxpublishsection" API action accepts an "issandbox" parameter, that
indicates whether a section translation should be published in the
user's sandbox instead of the main namespace. However, this parameter
is not passed to the API action for section translations on desktop,
leading to the parameter always be set to the default value of false,
and section translations always being published in the main namespace.

This patch fixes this issue by properly populating the value of the
"issandbox" parameter. It also prevents the target section title from
being wrongfully updated when the publishing target changes.

Bug: T316851
Change-Id: I5da0d6c6fa1df359dabcf9e6a5461d63ff67841d