CX: Fix transformation of TranslationUnitDTO to custom array
Currently, the output of the "contenttranslationcorpora" API endpoint,
doesn't include any parallel corpora content, missing the "mt", "user"
and "source" blobs are missing for all translation units.

This is caused by two issues:
1. The data of each translation unit are initially encapsulated inside
a TranslationUnitDTO object, then transformed into an associative
array, using the "toCustomArray" method of the TranslationUnitDTO
class. However, there is a typo inside that class that leads to errors.
2. The "toCustomArray" method of the TranslationUnitDTO expects an
array of valid types that are used for the given translation unit.
This array should be in the following form: ['user', 'source', 'mt'].
However, currently the array of types is wrongfully flipped leading,
and no valid blob type is found to be returned for the unit.

Both issues are fixed in this patch.

Bug: T326278
Change-Id: Ic639ffcfa6380d064676717212bc0b7c23eeb490