SX: Fix "getPage" getter to return null when title is empty
Until now, the "mediawiki/getPage" getter compares the given title with
the page title for each page inside the "mediawiki" state. If the titles
do not match, the getter checks if the page alias is not null and then
if the page alias is the same as the given title.

However, the page alias for a lot of pages inside the state, is
undefined (but not null). That means if the given title is also
undefined, the getter can match the undefined given title with a random
page that exists inside the state and happens to have an undefined
alias.

This has only recently become an issue for us, because we have updated
the code to also use the target page during publishing functionality.
That led to random pages being used as target pages when lead sections
are translated, because of the above described bug.

This patch fixes this issue by only checking the alias of a page when
it's truthy (thus undefined given titles won't match with any page).

Bug: T326566
Change-Id: Idc376663869b1df96e46c8b8e193a7f34d3bfbb1