SX: Fix MT parallel corpora to always use the correct MT service
Currently, when constructing the array of parallel corpora translation
units, for the "sxsave" payload, the MT provider used to translate
the first sentence of each subsection is used as origin for the MT
translation unit. That means that if the first sentence is not
translated (i.e. if it is skipped), the origin of the MT translation
unit will be undefined (since there is no MT provider used for the
first sentence). That leads the parallel corpora payload to fail the
validation and the auto-save request is not sent at all.

This patch fixes this issue by using the MT provider used for the
first TRANSLATED sentence (not just the first sentence inside the
subsection/paragraph) as origin for the MT translation unit.

Bug: T327080
Change-Id: Ic7f25b25fd850d00113447c617b47ad6f929c3e1