SX: Fix language selector for draft translations
The language selector inside "In progress" translations list doesn't
work properly. The users cannot really select the desired source or
target language for their translations. Even worse, the language pair
displayed for those draft translations, is always set to the source
and target languages of the application state, although all the draft
translations, regardless which their source/target languages are, are
displayed in the "In progress" list.

This patch fixes this issue by properly modifying the language selector
components (MWLanguageSelector, SXTranslationListLanguageSelector) to
support the "All languages" option, and then use the language selector
SFC properly inside the CXTranslationList component, which supports
the "In progress" translation list.

Bug: T327383
Change-Id: I942cc90692939ef10d822f07f4feb6f9bf321559