SX recent translation entrypoint: Fix query to use proper page title
Until now, the recent translation entrypoint was not showing, because
the query for fetching the related translation was using a wrong page
title. More specifically, it used "$out->getPageTitle()" method to
get the page title, which returns the HTML string to be placed inside
the H1 element of the page.

This patch fixes this issue by using the same method, we are using
during translation creation ($title->getPrefixedText()).

Bug: T316762
Change-Id: I9d2acb257dd35a52b06ce573c348f05119614a60