Make link parsing agnostic to baseURI, wgScript, wgArticlePath
A recent change in VE link parsing started using baseURI for link parsing.

The content provided by cxserver, whether it is article or machine
translated sections, has links with href="Title". But when this `href`
is parsed in the context of baseURI, it creates different URLs depending
on baseURI, wgScript, wgArticlePath.
For example, in a wiki where wgArticlePath is /wiki/$1, but baseURI based
on current url is /w/index.php?title=Title, the `href` must be /wiki/Title
for a valid URL. But cxserver cannot know this. And this can also happen
in local development wiki which follows different wgArticlePath than a production
wiki.
To complicate this further, CX always works with two language wikis.

For above mentioned factors, a valid internal mediawiki link may not be
recognized as internal link. This will result cxLink nodes throwing errors.

To fix this, have a defensive logic to construct the dataElement independent
of baseURI, wgScript, wgArticlePath.

Bug: T328143
Change-Id: I945897a27db479986855002b389034a745bf9bef