CX: Move "isTranslator" static method to TranslatorService
This patch is part of a group of patches that have as ultimate goal,
to move all functionality out of "Translator" and "Translation" classes.
This is done, because these classes act both as DTOs and as services,
violating the single responsibility principle and also introduce hidden
dependencies, because of the static methods they use.

Change-Id: I617242516fa45249b54fdcdd286f2c5cf5d70b32