Fix entrypoints to use page title instead of page name
"wgPageName" JS configuration parameter returns the page title
concatenated using "_", while "wgTitle" returns the page title
with spaces when the title consists of more than one word. For
the entrypoint URLs, we need to use the "wgTitle" parameter as
this is what CX and SX applications expect.

This patch fixes all entrypoints that used the wrong parameter
(2 in total).

Bug: T328939
Change-Id: I104de793a67dc23a15625a132fa8587d94aeb745