Allow suppressing SX invitation dialog via JS variable
Add a simple mechanism for skipping the section translation
invitation dialog for a single pageviews.

The immediate use case is GrowthExperiments post-edit notices
potentially conflicting with these dialogs. As the post-edit
notice is the more specific invitation (it's only shown right
after the Nth edit for some configuratble N), it makes sense
to prioritize that.

Using a JS variable means the ext.cx.entrypoints.recentedit
module will still get loaded but will be a no-op. That's not
ideal but there is no good way to do the suppressing on the
PHP side because the BeforePageDisplay hook would be the right
place both for another extension doing the suppressing and for
CX acknowledging the suppressing, and there is no way to
guarantee the other extension's BeforePageDisplay hook would
run first.

Bug: T322387
Change-Id: Ieda31508e7d6b03a833dbc7661e864f637e17830