SX: Support multiple translation work deletion at once
As demonstrated in the following screencast, deleting multiple
in-progress translations at once, leads to unexpected behaviour.
This happens because the key of each translation work inside the v-for
directive for the translation work list is based on its index inside
the array of translations. However, when the first translation is
deleted, these indexes are updated, leading to this unexpected result.

This patch fixes this issue by basing that key on the id of the
translation. This way, the keys are not updated after each translation
deletion and the deletion of multiple translation works at once is
possible.

Screencast: https://phab.wmfusercontent.org/file/data/nnvsy3qjf7jodsykwr2n/PHID-FILE-pmggbljqpzhb5mzbcyng/deleting_others_are_cleared.webm

Bug: T327385
Change-Id: Ia5b05ee3bebbe73f7671c489728fa906d85a0185