CX: Refactor parallel corpora related services
Code related to the parallel corpora is quite complicated, having two
different services ("CorporaLookup" and "TranslationCorporaStore")
that access the database and providing similar functionality (e.g.
fetching translation units from the database).

To clean up the code, this patch removes the "CorporaLookup" class
and moves all the database-access related functionality to the
"TranslationCorporaStore". At the same time, this patch adds a new
service, called "TranslationCorporaManager". This service relies on
the "TranslationCorporaStore" to fetch translation units from the
database and serves them in the several forms to its "clients". Such
clients are several Action API endpoints that need access to these
translation units, and some php scripts.

Change-Id: I9307d0a90a146b8382315af15837f8a1afc19ef5