Fix navigation from "Pick a sentence" step to dashboard
Currently, clicking the "back" icon inside "Pick a sentence" step leads
to the "Compare contents" step. However, the desired behaviour is that
the back icon should navigate the user back to the dashboard.

This patch supports this desired behaviour, by redirecting the user
back to the dashboard. Before completing the navigation, this patch
also sends a request to fetch the user translations, so that the newly
created draft inside "Pick a sentence" step, is also displayed in the
"In progress" dashboard translation list.

Bug: T287082
Change-Id: Iaa4dc9de489208c3c28bf338ffee4821d738602e