SX: Instrument follow-up invite shown after publishing
This patch modifies the component that displays the follow-up invite,
after successful publishing in SX, to use an anchor element instead
of redirecting through a click handler (for better accessibility) and
also adds the proper campaign to the URL, so that the invitation is
instrumented.

The patch also adds this new campaign in the list of supported event
sources for SX instrumentation.

Bug: T317995
Change-Id: I70ac9419f3bb7baf4da02c536c5f790f0cd8b875