SX: Add confirmation dialog on back button click
When users click on the back button inside "Pick a sentence" step,
a warning about losing current translation progress should be
displayed to the user. For this reason, this patch adds a dialog that
displays this warning to the user, with options to discard or continue
the current translation. This dialog is only displayed when a pending
"save translation" request exists, otherwise the user is directly
navigated to the dashboard.

Bug: T270484
Change-Id: I5c69fda97c6d3d910c7b11af7b7758b5f06e4d8b