SX: Remove "autoSaveInProgressCounter" variable from application state
"autoSaveInProgressCounter" state variable seems redundant since the
"autoSavePending" state variable will always be true, when the counter
variable is a positive integer. The only case, that this is not true,
i.e. the only case when "autoSavePending" variable is false and the
"autoSaveInProgressCounter" is greater than zero, is when the
"publishTranslation" action is executed. However, we can simply ignore
this case, since the user cannot really go to the dashboard, when
publishing is pending.

For these reasons, this patch removes the "autoSaveInProgressCounter"
from the application state, to avoid polluting the state with
redundant variables.

Change-Id: I315eb3ea2b283a2df5374b4903b7b438b797fc94