SX: Add "reset" methods to page section, subsection and sentence models
Current translation should be reset, when user leaves the "Pick a
sentence" step and returns to the dashboard. This patch adds the
needed methods inside the page section, subsection and sentence
models, so that the contents of the translation can be properly
reset. These methods will be used in a followup patch.

Bug: T333849
Change-Id: I530a9993564d79e66050d39c663c1967303be4d6