SX: Reset translation before going back to dashboard
The current translation should be reset when user returns to the
dashboard, so that in case they select the same article/section
for translation (without reloading the page), they should always
get a fresh section to translate. This patch uses the previously
added methods and actions to reset the translation for these cases.

Bug: T333849
Change-Id: Ia4e05bf116088797e2da0fb0f722f5b77412b252