SX instrumentation: Always take "campaign" URL parameter into account
Currently, the "campaign" URL parameter is taking into account, only
when the "page" and "sx" URL parameters are also set. However, that
leads to instrumentation events being logged with wrong event_source
(always set to "direct") when no page is provided. This patch fixes
this issue by using the proper event_source, even when no page title
is provided in the URL.

The motive for this change, is that the persistent entrypoint is not
properly logged in SX, because of the above issue.

Bug: T321386
Change-Id: I32e1844ebde34a32a33b29d435b702fb6f080286