SX draft restoration: Continue translation from where it was stopped
Currently, when restoring a translation on SX, the title is always
initially selected, even if the user has translated a lot of content
of the section.

This patch fixes this issue by setting as active, the sentence which
is next to the last translated sentence. That is if sentences 1,2,5
have a translation, the active sentence will be 6.

Bug: T327495
Change-Id: I9dd774689ad11527da3b5d44478eafd7094694ad