SX: Reset current translation, if any, when going back to dashboard
The motive for this change is as I noticed, if a user loads a draft
translation, then goes back to dashboard and picks a new section
translation to translate, the selected segment inside "Pick a sentence"
step is always the segment next to the title, instead of the title.
This is caused by the non-empty "currentTranslation" state variable.

Change-Id: I0d1b3b96e35f7ac68e0ea9fc56bb56e8d04f566c