SX: Move template rendering functionality to service
The HTML of block templates as returned both by cxserver and by Visual
Editor surface "getHTML" method, only contain the definition of the
block template. For that reason, we need to extract the template
definition from those HTML strings as wikitext and transform it back to
HTML by sending a request to the "visualeditor" API action, in order to
render the template properly in the UI. Until now, this logic lived
inside vuex actions, however since the same functionality will be
needed in order to adapt the HTML of restored templates (for draft
translations), it is important to move them to a separate service.

This will enable us to reuse code, reason about this code, debug and
test it more easily. This is why, this patch moves this functionality
to a separate file named "templateRenderer".

At the same time, the "setBlockTemplateAdaptationInfo" method inside
"subSection" model, is renamed to "setBlockTemplateAdaptationInfoByHtml"
and expanded to also handle the prerequisite checks for the block
template HTML, before setting the "blockTemplateAdaptationInfo" class
variable.

Change-Id: I1b68f8183693c95d608d42a2664d69c751963583