SX: Support restoration of block templates for draft translations
Until now when a draft translation was restored, the translated
block templates were not restored at all, the original template
was displayed instead of the translated one.

This patch adds support for the restoration of block templates,
by properly handling the re-rendering of the template translation,
inside the "restoreBlockTemplate" method of the translationRestorer
service.

Additionally, it modifies the corpora payload for block templates
in SX to also include the template definition (data-cx) attribute,
so that templates translated in SX can be properly restored. This
is not necessary for templates translated in CX, as this attribute
is already present.

The patch also adds some simple modifications to the template HTML
for translated templates coming from CX, as there are some small
differences between the subsection wrapper saved in CX in comparison
to the template wrappers for templates translated in SX.

Finally, the patch waits for all the template restorations to be
completed, before the current translation is considered "restored".
This waiting was not necessary until now, as the restoration of
"text" subsections is synchronous. However, the restoration of
block templates is asynchronous and we have to wait for it.

Bug: T336149
Change-Id: I6c48c39934639de99c9a9ec491a59e91f897646b